Code/data_analysis/requirements.txt
paho-mqtt==1.6.1
requests==2.26.0
numpy==1.21.0